requirements.txt
asgiref==3.7.2
Babel==2.13.0
beautifulsoup4==4.12.2
crispy-bootstrap5==0.7
Django==4.2.6
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
django-money==3.3.0
psycopg2-binary==2.9.9
py-moneyed==3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
